+++ spec/rubocop/cop/lint/block_alignment_spec.rb
@@ -157,19 +157,19 @@ describe Rubocop::Cop::Lint::BlockAlignment do
     # Example from issue 447 of bbatsov/rubocop on github:
     it 'accepts two kinds of end alignment' do
       src = [
-             # Aligned with start of line where do is:
-             'params = default_options.merge(options)',
-             '          .delete_if { |k, v| v.nil? }',
-             '          .each_with_object({}) do |(k, v), new_hash|',
-             '            new_hash[k.to_s] = v.to_s',
-             '          end',
-             # Aligned with start of the whole expression:
-             'params = default_options.merge(options)',
-             '          .delete_if { |k, v| v.nil? }',
-             '          .each_with_object({}) do |(k, v), new_hash|',
-             '            new_hash[k.to_s] = v.to_s',
-             'end'
-            ]
+        # Aligned with start of line where do is:
+        'params = default_options.merge(options)',
+        '          .delete_if { |k, v| v.nil? }',
+        '          .each_with_object({}) do |(k, v), new_hash|',
+        '            new_hash[k.to_s] = v.to_s',
+        '          end',
+        # Aligned with start of the whole expression:
+        'params = default_options.merge(options)',
+        '          .delete_if { |k, v| v.nil? }',
+        '          .each_with_object({}) do |(k, v), new_hash|',
+        '            new_hash[k.to_s] = v.to_s',
+        'end'
+      ]
       inspect_source(cop, src)
       expect(cop.offenses).to be_empty
     end